% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.yc_curve}
\alias{plot.yc_curve}
\title{Plot Method for Yield Curve Objects}
\usage{
\method{plot}{yc_curve}(x, ...)
}
\arguments{
\item{x}{A \code{yc_curve} object.}

\item{...}{Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
The input object, invisibly.
}
\description{
Plot Method for Yield Curve Objects
}
