% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot.yc_pca}
\alias{plot.yc_pca}
\title{Plot Method for Yield Curve PCA Objects}
\usage{
\method{plot}{yc_pca}(x, ...)
}
\arguments{
\item{x}{A \code{yc_pca} object.}

\item{...}{Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
The input object, invisibly.
}
\description{
Plots the factor loadings for each principal component across tenors.
}
