% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.yc_curve}
\alias{print.yc_curve}
\title{Print Method for Yield Curve Objects}
\usage{
\method{print}{yc_curve}(x, ...)
}
\arguments{
\item{x}{A \code{yc_curve} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Yield Curve Objects
}
