% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{print.yc_pca}
\alias{print.yc_pca}
\title{Print Method for Yield Curve PCA Objects}
\usage{
\method{print}{yc_pca}(x, ...)
}
\arguments{
\item{x}{A \code{yc_pca} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Yield Curve PCA Objects
}
