% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.yc_curve}
\alias{summary.yc_curve}
\title{Summary Method for Yield Curve Objects}
\usage{
\method{summary}{yc_curve}(object, ...)
}
\arguments{
\item{object}{A \code{yc_curve} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Summary Method for Yield Curve Objects
}
