% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{summary.yc_pca}
\alias{summary.yc_pca}
\title{Summary Method for Yield Curve PCA Objects}
\usage{
\method{summary}{yc_pca}(object, ...)
}
\arguments{
\item{object}{A \code{yc_pca} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Summary Method for Yield Curve PCA Objects
}
